requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
aiogram==3.22.0
pydantic==2.11.10
python-multipart==0.0.6
asyncpg==0.29.0
aiosqlite==0.19.0
python-dotenv==1.0.0
